==> u02/02_predictions.ipynb <==
import tensorflow as tf
from tensorflow.keras.models import load_model
import pickle
import pandas as pd
import numpy as np

### Load the pickle file (trained model, scalar, onehot)
model = load_model('model.h5')

with open('onehot_encoder_geo.pkl', 'rb') as file:
    label_encoder_geo = pickle.load(file)

with open('label_encoder_gender.pkl', 'rb') as file:
    label_encoder_gender = pickle.load(file)

with open('scaler.pkl', 'rb') as file:
    scaler = pickle.load(file)

# Example input data to do the prediction
input_data = {
    'CreditScore': 600,
    'Geography': 'France',
    'Gender': 'Male',
    'Age': 40,
    'Tenure': 3,
    'Balance':60000,
    'NumOfProducts': 2,
    'HasCrCard': 1,
    'IsActiveMember': 1,
    'EstimatedSalary': 50000
}
print(input_data)

geo_encoded=label_encoder_geo.transform([[input_data['Geography']]]).toarray()
print(input_data['Geography'])
print(geo_encoded)
geo_encoded_df=pd.DataFrame(geo_encoded,columns=label_encoder_geo.get_feature_names_out(['Geography']))
geo_encoded_df

input_df=pd.DataFrame([input_data])
input_df

# Combine one-hot encoded columns with input data
# row concatenation - axis=1
input_df = pd.concat([input_df.reset_index(drop=True), geo_encoded_df], axis=1)
input_df

# Encode categorical variables
input_df['Gender'] = label_encoder_gender.transform(input_df['Gender'])
input_df

# Scaling the input data
input_scaled=scaler.transform(input_df)
input_scaled

## Predict churn
prediction=model.predict(input_scaled)
prediction

prediction_proba=prediction[0][0]
prediction_proba

if prediction_proba > 0.5:
    print("the customer is likely to churn.")
else:
    print("the customer is not likely yo churn")
