==> u02/04_regression.ipynb <==
import tensorflow as tf
import pickle
import pandas as pd
import numpy as np
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import StandardScaler, LabelEncoder
from sklearn.model_selection import train_test_split

data=pd.read_csv('Churn_Modelling.csv')
data

# Oreprocess the data
data=data.drop(['RowNumber','CustomerId','Surname'], axis=1)
data

# One hot encode 'Geography'
onehot_encoder_geo=OneHotEncoder(handle_unknown='ignore')
geo_encoded = onehot_encoder_geo.fit_transform(data[['Geography']]).toarray()
geo_encoded_df = pd.DataFrame(geo_encoded, columns=onehot_encoder_geo.get_feature_names_out(['Geography']))
geo_encoded_df

## Combine one hot encoder columns with the original data
data=pd.concat([data.drop('Geography',axis=1),geo_encoded_df],axis=1)
data.head()

## Encode categorical variables
label_encoder_gender=LabelEncoder()
data['Gender']=label_encoder_gender.fit_transform(data['Gender'])
data

# Split the data into features and target
X = data.drop('EstimatedSalary', axis=1)
y = data['EstimatedSalary']

## Split the data in training and tetsing sets
print(X.shape)
print(y.shape)
X_train,X_test,y_train,y_test=train_test_split(X,y,test_size=0.2,random_state=42)

print(X_train.shape)
print(y_train.shape)

## Scale these features
scaler=StandardScaler()
X_train=scaler.fit_transform(X_train)
X_test=scaler.transform(X_test)

X_train

y_test

## Save the encoders and sscaler
# Pickling” is the process whereby a Python object hierarchy is converted into a byte stream
# here we are saving to a file the obects: label_encoder_gender, onehot_encoder_geo
#with open('label_encoder_gender.pkl','wb') as file:
#    pickle.dump(label_encoder_gender,file)

#with open('onehot_encoder_geo.pkl','wb') as file:
#    pickle.dump(onehot_encoder_geo,file)

#with open('scaler.pkl','wb') as file:
#    pickle.dump(scaler,file)


## ANN Regression Problem statement

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense

print(X_train.shape[0]) # rows
print(X_train.shape[1]) # columns

# Build the model
# the neural tetwork with thte 2 hid
model=Sequential([
    Dense(64,activation='relu',input_shape=(X_train.shape[1],)), ## HL1 Connected wwith input layer, 64 is decided initially
    Dense(32,activation='relu'), ## HL2
    Dense(1,activation='sigmoid')  ## output layer
])

# compile the model
# https://keras.io/api/losses/
model.compile(optimizer='adam', loss='mean_absolute_error', metrics=['mae'])

model.summary()

from tensorflow.keras.callbacks import EarlyStopping, TensorBoard
import datetime

# Setup TensorBoard
log_dir="regressionlogs/fit/"+datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
tensorboard_callback=TensorBoard(log_dir=log_dir, histogram_freq=1)

# Set up Early Stopping
early_stopping_callback=EarlyStopping(monitor='val_loss',patience=10, restore_best_weights=True)

# Train the model
history=model.fit(
    X_train, y_train,
    validation_data=(X_test,y_test),
    epochs=100,
    callbacks=[early_stopping_callback, tensorboard_callback]

)

%load_ext tensorboard

%tensorboard --logdir regressionlogs/fit
#20240912-205012


# Evaluate model on the test data
test_loss, test_mae=model.evaluate(X_test,y_test)
print(f'Test MAE: {test_mae}')

model.save('regression_model.h5')
