==> u02/06_embeddings.ipynb <==
from tensorflow.keras.preprocessing.text import one_hot

sent = ['the glass of milk',
        'the glass of juice',
        'the cup of tea',
        'I am a good boy',
        'I am a good developer',
        'understand the meaning of words',
        'your videos are good']

# Define the vocabulary size
voc_size = 10000

# One hot representation
one_hot_repr=[one_hot(word,voc_size) for word in sent]
one_hot_repr

# word embedding representation
from tensorflow.keras.layers import Embedding
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.models import Sequential

import numpy as np

# padding to have sentences of the same lenght

sent_length=8 # rounded up from the number of words in the list of sentences above
embedded_docs=pad_sequences(one_hot_repr,padding='pre',maxlen=sent_length)
print(embedded_docs)

# feature representation
dim=10

# features are the possible classification of each word (ex. male, age, food, ...)s
model=Sequential()
model.add(Embedding(voc_size,dim,input_length=sent_length))
model.compile('adam','mse')

print(f"sent_length: {sent_length}, voc_size: {voc_size}")

model.summary()

model.predict(embedded_docs)
# each word is represented by 10 items in the array

# 8 words per sentence
# 10 features
# ->
# 8 subarray (one per word in the sentence)
# 10 is the lenght of each subarray (one per feature)
model.predict(embedded_docs[0])

embedded_docs[0].shape
