==> u02/07_simplernn.ipynb <==
End to end Deep Learning Project using simple RNN

import numpy as np
import tensorflow as tf
from tensorflow.keras.datasets import imdb
from tensorflow.keras.preprocessing import sequence
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Embedding, SimpleRNN, Dense

# Load the imdb dataset
max_features=10000 # vocabulary size
(X_train, y_train),(X_test,y_test)=imdb.load_data(num_words=max_features)

# print the share of the data
print(f'Training data shape: {X_train.shape}, Training labels shape: {y_train.shape}')
print(f'Testing data shape: {X_train.shape}, Testing labels shape: {y_train.shape}')

# explore the imdb
sample_review=X_train[0]
sample_label=y_train[0]

X_train[0] # this is an array that represents one sentence

# output can be 1 or 0 
# because we have sentiment analysis 
# 0 means negative sentiment
# 1 means positive sentiment
y_train[0]

print(f'Sample review (as integer):{sample_review}')
print(f'Sample label:{sample_label}')

#Mapping index back to words
word_index=imdb.get_word_index()
word_index

# use the key as index of the dictionary
index_from=3 # this is how the dataset is done... refer to keras documentation
reverse_word_index={i+index_from:word for word, i in word_index.items()}
reverse_word_index

# decode the array of the sample sentence
# from the word index
decoded_review=' '.join(reverse_word_index.get(i,'?') for i in sample_review)
print(decoded_review)

from tensorflow.keras.preprocessing import sequence
max_len=500

X_train=sequence.pad_sequences(X_train, maxlen=max_len)
X_test=sequence.pad_sequences(X_test, maxlen=max_len)
X_train

## Train Simple RNN

model=Sequential()
model.add(Embedding(max_features, 128, input_length=max_len)) # embedding layer
model.add(SimpleRNN(128, activation='relu')) # 128 neurons
model.add(Dense(1,activation='sigmoid')) # output

model.summary()

model.compile(optimizer='adam',loss='binary_crossentropy',metrics=['accuracy'])

# create an instance of EarlyStopping Callback
from tensorflow.keras.callbacks import EarlyStopping
earlystopping=EarlyStopping(monitor='val_loss', patience=5, restore_best_weights=True)

# Train the model with ealy stopping
history=model.fit(
    X_train,y_train,epochs=10,batch_size=32,
    validation_split=0.2,
    callbacks=[earlystopping]
)


# save model file
model.save('simple_rnn_imdb.h5')
