==> u02/08_prediction.ipynb <==
import numpy as np
import tensorflow as tf
from tensorflow.keras.datasets import imdb
from tensorflow.keras.preprocessing import sequence
from tensorflow.keras.models import Sequential, load_model
from tensorflow.keras.layers import Embedding, SimpleRNN, Dense

#Mapping index back to words
word_index=imdb.get_word_index()
# use the key as index of the dictionary
index_from=3 # this is how the dataset is done... refer to keras documentation
reverse_word_index={i+index_from:word for word, i in word_index.items()}
reverse_word_index

model=load_model('simple_rnn_imdb.h5')
model.summary()

model.get_weights()

(X_train, y_train),(X_test,y_test)=imdb.load_data(num_words=10000)

# Step 2: Helper functions
# functions to decode reviews
def decode_review(encoded_review):
    return ' '.join([reverse_word_index.get(i,'?') for i in encoded_review])

# function to preprocess user input
def preprocess_text(text):
    words=text.lower().split()
    encoded_review=[word_index.get(word,2) for word in words]
    padded_review = sequence.pad_sequences([encoded_review], maxlen=500)
    return padded_review


decode_review(X_train[0])

# prediction function
def predict_sentiment(review):
    preprocessed_input=preprocess_text(review)
    prediction=model.predict(preprocessed_input)
    sentiment = 'Positive' if prediction[0][0] > 0.5 else 'Negative'
    return sentiment, prediction[0][0]

example_review="This movie was fantastic! The acting was great and plot was thrilling."

sentiment,score=predict_sentiment(example_review)

print(f'Review: {example_review}')
print(f'Sentiment: {sentiment}')
print(f'Prediction score: {score}')
